Small changes made in all three classes
Changed getBoard() in Grid class to check for out of bounds, but should it
throw an exception or return null when outside bounds? Also, Grid class
now has get/set whoIsWinning() methods for convenience. Added a variable
'turn' with type Cell in Board and Grid classes to determine whose turn it
is. Also, the setCell() method in class Board does not contain the mark
parameter but rather uses the turn variable to determine which Cell is
placed on the board. Cell class is now based on color because we are not
using X's and O's to represent the mark.