The final intermediate submission
I did minimal testing on the GUI, I'm not sure how to test all of the
elements such as the dialogFragment or the actionbar items. I also made
the winning shade less opaque so that the player can see just how they
won/loss more easily and I added a toast to go along with that shade.

There is a bug where you can still play after the game is won, I've
alerted Charlie of this bug.