Added an animation for the last move
Also added a visual cue for the last move, it is more opaque than all of
the other moves. This required me to change the opacity of various shapes,
but I think I got a good balance on them all (basically making everything
except the last move more transparent so that the completely opaque last
move is more obvious).